Revert "Added solindex cron"
This reverts commit 01c4009211c3c4338d7edec6ffe8c2de62496f75.

+++ solindex_cron/app.py
@@ -5,7 +5,6 @@ from datetime import datetime
 import boto3
 from botocore.exceptions import ClientError
 import os
-from decimal import Decimal
 
 
 def fetch_data(url):
@@ -37,7 +36,7 @@ def parse_html(html):
                      .replace('₉', '9')
                      )
             try:
-                price = Decimal(price)
+                price = float(price)
                 data.append((name, price))
             except ValueError:
                 print(f"Skipping invalid price: {price}")
